// src/youtube_music/types/ExploreResults_0.ts
export interface ExploreResults_0 {
    contents: Content[]
    trackingParams: string
}

export interface Content {
    gridRenderer?: GridRenderer
    musicCarouselShelfRenderer?: MusicCarouselShelfRenderer
}

export interface GridRenderer {
    items: Item[]
    trackingParams: string
}
// ...
export interface MusicCarouselShelfRenderer {
    header: Header
    contents: Content2[]
    trackingParams: string
    itemSize: string
    numItemsPerColumn?: string
}