// src/youtube_music/types/LibraryResults_0.ts
export interface ParsedLibraryResults {
    "title": string,
    "thumbnail_uri": string,
    "endpoint": string
}
export interface LibraryResults_0 {
    responseContext: ResponseContext
    contents: Contents
    trackingParams: string
    frameworkUpdates: FrameworkUpdates
}

export interface ResponseContext {
    serviceTrackingParams: ServiceTrackingParam[]
}

export interface ServiceTrackingParam {
    service: string
    params: Param[]
}

export interface Param {
    key: string
    value: string
}

export interface Contents {
    singleColumnBrowseResultsRenderer: SingleColumnBrowseResultsRenderer
}

export interface SingleColumnBrowseResultsRenderer {
    tabs: Tab[]
}
// ...
export interface FrameworkUpdates {
    entityBatchUpdate: EntityBatchUpdate
}

export interface EntityBatchUpdate {
    mutations: Mutation[]
    timestamp: Timestamp
}