// src/youtube_music/types/SearchResults_0.ts
export type SearchResults_0 = Root2[]

export interface Root2 {
    responseContext: ResponseContext
    items?: Item[]
    trackingParams: string
    contents: Contents
}

export interface ResponseContext {
    serviceTrackingParams: ServiceTrackingParam[]
    maxAgeSeconds?: number
}

export interface ServiceTrackingParam {
    service: string
    params: Param[]
}

export interface Param {
    key: string
    value: string
}

export interface Item {
    guideSectionRenderer: GuideSectionRenderer
}
// ...
export interface Contents {
    tabbedSearchResultsRenderer: TabbedSearchResultsRenderer
}